opapremodit and manual agreement test separated
