+++ tests/unittests/spec/premodit/premodit_modit_agreement_test.py
@@ -2,21 +2,14 @@
 """
 
 import pytest
-import numpy as np
 import jax.numpy as jnp
 from exojax.spec.opacalc import OpaPremodit
-from exojax.spec.premodit import unbiased_lsd_zeroth
-from exojax.spec.premodit import unbiased_lsd_first
-from exojax.spec.premodit import unbiased_lsd_second
-from exojax.spec.lsd import inc2D_givenx
 from exojax.spec.hitran import line_strength
 from exojax.spec.set_ditgrid import ditgrid_log_interval
 from exojax.spec.exomol import gamma_exomol
 from exojax.utils.constants import Tref_original
 from exojax.test.emulate_mdb import mock_mdbExomol
 from exojax.test.emulate_mdb import mock_wavenumber_grid
-from exojax.spec.modit_scanfft import calc_xsection_from_lsd_scanfft
-from exojax.spec.premodit import unbiased_ngamma_grid
 from exojax.spec import normalized_doppler_sigma
 from exojax.spec import initspec
 
@@ -77,71 +70,7 @@ def test_agreement_premodit_modit(diffmode):
     )  # maximum differnce =  0.0058, 0.004, 0.008, for diffmode=0,1,2 2/4 2025 @manbou
 
 
-@pytest.mark.parametrize("diffmode", [0, 1, 2])
-def test_premodit_opa_and_manual_agreement(diffmode):
-    """Test the agreement between the automatic computation of cross section by xsvector 
-    and manual computation of cross section by calc_xsection_from_lsd_scanfft. for Premodit
-    """
-    from jax import config
-
-    config.update("jax_enable_x64", True)
-
-    mdb = mock_mdbExomol()
-    nus, wav, res = mock_wavenumber_grid()
-    Ttest = 1200.0
-    P = 1.0
-    # PreMODIT LSD
-    opa = OpaPremodit(
-        mdb=mdb, nu_grid=nus, auto_trange=[1000.0, 1500.0], diffmode=diffmode
-    )
-    (
-        lbd_coeff,
-        multi_index_uniqgrid,
-        elower_grid,
-        ngamma_ref_grid,
-        n_Texp_grid,
-        R,
-        pmarray,
-    ) = opa.opainfo
-
-    # automatic computing by xsvector
-    xsv = opa.xsvector(Ttest, P)
-    # tries manual computation of xsvector below
-    qt = mdb.qr_interp(Ttest, opa.Tref)
-    if diffmode == 0:
-        Slsd_premodit = unbiased_lsd_zeroth(
-            lbd_coeff[0], Ttest, opa.Tref, nus, elower_grid, qt
-        )
-    elif diffmode == 1:
-        Slsd_premodit = unbiased_lsd_first(
-            lbd_coeff, Ttest, opa.Tref, opa.Twt, opa.nu_grid, elower_grid, qt
-        )
-    elif diffmode == 2:
-        Slsd_premodit = unbiased_lsd_second(
-            lbd_coeff, Ttest, opa.Tref, opa.Twt, opa.nu_grid, elower_grid, qt
-        )
-    nsigmaD = normalized_doppler_sigma(Ttest, mdb.molmass, R)
-    ngamma_grid = unbiased_ngamma_grid(
-        Ttest,
-        P,
-        ngamma_ref_grid,
-        n_Texp_grid,
-        multi_index_uniqgrid,
-        opa.Tref_broadening,
-    )
-    log_ngammaL_grid = jnp.log(ngamma_grid)
-    xsv_manual = calc_xsection_from_lsd_scanfft(
-        Slsd_premodit, R, pmarray, nsigmaD, nus, log_ngammaL_grid
-    )
-
-    dxsv = jnp.abs(xsv_manual / xsv - 1)
-    assert np.max(dxsv) < 1.0e-11  # (several x e-12, feb 4th 2025)
-
-
 if __name__ == "__main__":
-    test_premodit_opa_and_manual_agreement(diffmode=0)
-    test_premodit_opa_and_manual_agreement(diffmode=1)
-    test_premodit_opa_and_manual_agreement(diffmode=2)
     test_agreement_premodit_modit(diffmode=0)
     test_agreement_premodit_modit(diffmode=1)
     test_agreement_premodit_modit(diffmode=2)
